Remove an extra_require setting that breaks `pants generate-lockfile`
Pants versions 2.19 and 2.20 choke on this with error:

```
11:10:45.65 [INFO] Completed: Generate lockfile for python-default
11:10:45.65 [ERROR] 1 Exception encountered:

Engine traceback:
  in `generate-lockfiles` goal

ProcessExecutionFailure: Process 'Generate lockfile for python-default' failed with exit code 1.
stdout:

stderr:
Expected matching RIGHT_PARENTHESIS for LEFT_PARENTHESIS, after version specifier
    cadquery (==master) ; extra == 'prod'
             ~^
```

A standalone reproducer for the issue is available at
https://github.com/eguiraud-pf/pants-219-lockfile-issue-repro.
It was reported to the pants dev but no solution was proposed, and since
the jupyter-cadquery package is not being actively developed since 2-3
years it should be safe to fork it and patch it instead.